+++ SplunkAgent/Sources/SplunkAgent/Public API/API-1.0-RuntimeState.swift
@@ -68,9 +68,9 @@ extension RuntimeState {
         owner.agentConfiguration.appVersion
     }
 
-    /// A ``EndpointConfiguration`` containing either the specified realm, or endpoint urls.
-    public var endpointConfiguration: EndpointConfiguration {
-        owner.agentConfiguration.endpoint
+    /// An optional ``EndpointConfiguration`` containing either the specified realm, or endpoint urls.
+    public var endpointConfiguration: EndpointConfiguration? {
+        owner.currentEndpoint
     }
 
     /// A `String` containing the used application deployment environment.